stop reading from cache

+++ main.rb
@@ -14,7 +14,7 @@ def fetch_rss_feed(url)
     FileUtils.mkdir( tmp_dir )
   end
   location = File.join(tmp_dir, "#{host}.xml")
-  if File.exists?(location)
+  if File.exists?(location) && false
     puts "READING FROM FILE"
     feed = File.open(location).read
   else
